+++ src/main/java/tn/esprit/tp1_benelbeyskander_4twin7/reposetries/IresevationRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.tp1_benelbeyskander_4twin7.reposetries;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.tp1_benelbeyskander_4twin7.entities.Reservation;
 
-public interface IresevationRepository extends CrudRepository<Reservation,Integer> {
+public interface IresevationRepository extends CrudRepository<Reservation,Long> {
 }
