+++ src/main/java/tn/esprit/tp1_benelbeyskander_4twin7/services/IChmabreServices.java
@@ -7,7 +7,7 @@ public interface IChmabreServices {
     List<Chambre> getAllChmabres();
     Chambre addChambre(Chambre chambre );
         Chambre updateChambre(Chambre chambre);
-        Chambre getChambre(int idChambre);
-        void supprimerChambre(int idChambre);
+        Chambre getChambre(long  idChambre);
+        void supprimerChambre(long  idChambre);
 
 }
